pyproject.toml
[project]
name = "orbbec-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "customtkinter>=5.2.2",
    "numpy>=2.2.6",
    "open3d>=0.19.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyorbbecsdk>=1.3.2",
]
